kafka/server: Move request_types to seperate Hdr
- So that other features that must know of all request types can consume
from this master list
